Added Retrofit service for testing

+++ app/src/main/java/com/example/final_weather/models/CurrentWeather.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class CurrentWeather(
     @SerializedName("temp_c")
-    val currentC:Int
+    val currentC:Float
 
 )
